Re-set the default number of not_bias_corrected channels (#21)
Per @emilyhcliu 's discovery that the default number of "-999" for
not_bias_corrected channels confused UFO weeks ago. The reason is that
the dash sign "-" is configured in channels numbers such as "10-20" in
UFO, and then wrong channel indices can be returned when "-999" is
searched in UFO. Propose to re-set the default number as "99999".
@emilyhcliu 's tests confirmed that it could fix the issue.

Co-authored-by: Dan Holdaway <27729500+danholdaway@users.noreply.github.com>
Co-authored-by: Cory Martin <cory.r.martin@noaa.gov>